TASK: adding admin dashboard

+++ routes/api.php
@@ -22,14 +22,14 @@ Route::post('login', 'AuthController@login');
 
 Route::get('user', 'AuthController@user')->middleware('auth:api');
 
-Route::get('meetings', 'MeetingController@index');
-Route::post('meeting', 'MeetingController@store');
-Route::put('meeting/{id}', 'MeetingController@update');
-Route::delete('meeting/{id}', 'MeetingController@delete');
+Route::get('meetings', 'MeetingController@index')->middleware('auth:api');
+Route::post('meeting', 'MeetingController@store')->middleware('auth:api');
+Route::put('meeting/{id}', 'MeetingController@update')->middleware('auth:api');
+Route::delete('meeting/{id}', 'MeetingController@delete')->middleware('auth:api');
 
-Route::get('related_meetings/{user_id}', 'MeetingUserController@getRelatedMeetings');
-Route::get('pending_meetings/{user_id}', 'MeetingUserController@getPendingMeetings');
-Route::get('past_meetings/{user_id}', 'MeetingUserController@getPastMeetings');
-Route::get('registered_meetings/{user_id}', 'MeetingUserController@getRegisteredMeetings');
+Route::get('related_meetings/{user_id}', 'MeetingUserController@getRelatedMeetings')->middleware('auth:api');
+Route::get('pending_meetings/{user_id}', 'MeetingUserController@getPendingMeetings')->middleware('auth:api');
+Route::get('past_meetings/{user_id}', 'MeetingUserController@getPastMeetings')->middleware('auth:api');
+Route::get('registered_meetings/{user_id}', 'MeetingUserController@getRegisteredMeetings')->middleware('auth:api');
 
-Route::put('register_user/{user_id}_{meeting_id}', 'MeetingUserController@registerUser');
+Route::put('register_user/{user_id}_{meeting_id}', 'MeetingUserController@registerUser')->middleware('auth:api');
